02-31-Refactoring to make Testing Damaging Unit Possible

+++ AAR-GEMASTIK-2024/Assets/Scripts/Enemy/EnemyHealthSystem.cs
@@ -5,7 +5,7 @@ public struct UpdateEnemyHealthValue
 {
     public int currentValue;
     public int maxValue;
-    public readonly float PercentageValue => currentValue / maxValue;
+    public readonly float PercentageValue => (float)currentValue / maxValue;
     public UpdateEnemyHealthValue(int currentValue, int maxValue)
     {
         this.currentValue = currentValue;
